pyproject.toml
[project]
name = "hkjc-scraper"
version = "0.1.0"
description = "HKJC Racing Scraper - Extract horse racing data from Hong Kong Jockey Club"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiolimiter>=1.2.1",
    "psycopg2-binary>=2.9",
    "scrapling[all]>=0.4.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "beautifulsoup4>=4.12",
]

[project.scripts]
hkjc-scrape = "hkjc_scraper.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hkjc_scraper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (makes network requests)",
]
filterwarnings = [
    "ignore::DeprecationWarning:lxml.html",
]
